cody-gateway: clean up layers of limiters
