+++ enterprise/cmd/cody-gateway/internal/actor/BUILD.bazel
@@ -5,6 +5,7 @@ go_library(
     name = "actor",
     srcs = [
         "actor.go",
+        "limits.go",
         "source.go",
     ],
     importpath = "github.com/sourcegraph/sourcegraph/enterprise/cmd/cody-gateway/internal/actor",
@@ -30,7 +31,7 @@ go_library(
 go_test(
     name = "actor_test",
     srcs = [
-        "actor_test.go",
+        "limits_test.go",
         "source_test.go",
     ],
     embed = [":actor"],
